+++ src/commands/init.rs
@@ -19,14 +19,13 @@ use super::new::New;
 use crate::commands::helps::init;
 use crate::commands::orbit::AnyResult;
 use crate::core::context::Context;
-use crate::core::ip::Ip;
 use crate::core::lang::vhdl::token::Identifier;
 use crate::core::manifest::{Manifest, IP_MANIFEST_FILE};
 use crate::core::pkgid::PkgPart;
 use crate::error::{Error, LastError};
 use crate::util::anyerror::AnyError;
+use crate::util::filesystem;
 use crate::util::filesystem::Standardize;
-use crate::util::filesystem::{self, ORBIT_IGNORE_FILE};
 use std::io::Write;
 use std::path::PathBuf;
 
@@ -52,7 +51,7 @@ impl Subcommand<Context> for Init {
         })
     }
 
-    fn execute(self, c: &Context) -> proc::Result {
+    fn execute(self, _: &Context) -> proc::Result {
         // TODO: verify the pkgid is not taken
 
         // TODO: refactor due to heavy overlap with 'new' command
@@ -68,7 +67,7 @@ impl Subcommand<Context> for Init {
 
         let ip_name = New::extract_name(self.name.as_ref(), &dest)?;
 
-        match self.create_ip(&ip_name, &c.get_target_dir()) {
+        match self.create_ip(&ip_name) {
             Ok(r) => Ok(r),
             Err(e) => Err(Error::FailedToInitIp(LastError(e.to_string())))?,
         }
@@ -77,7 +76,7 @@ impl Subcommand<Context> for Init {
 
 impl Init {
     /// Initializes a project at an exising path.
-    fn create_ip(&self, ip: &PkgPart, target_dir: &str) -> AnyResult<()> {
+    fn create_ip(&self, ip: &PkgPart) -> AnyResult<()> {
         // verify the directory already exists
         if self.path.is_dir() == false || self.path.exists() == false {
             return Err(Box::new(AnyError(format!(
@@ -93,21 +92,13 @@ impl Init {
             p
         };
 
-        let ignore_path = {
-            let mut p = self.path.clone();
-            p.push(ORBIT_IGNORE_FILE);
-            p
-        };
-
         let lib_str = match &self.library {
             Some(s) => Some(s.to_string()),
             None => None,
         };
 
         let mut manifest = std::fs::File::create(&manifest_path)?;
-        let mut ignore = std::fs::File::create(&ignore_path)?;
         manifest.write_all(Manifest::write_empty_manifest(&ip, &lib_str).as_bytes())?;
-        ignore.write_all(Ip::write_default_ignore_file(target_dir).as_bytes())?;
         Ok(())
     }
 }