+++ tests/t7/target/ao/d.v
@@ -1,0 +1,2 @@
+module d;
+endmodule