+++ tests/t7/target/c.v
@@ -1,0 +1,2 @@
+module c;
+endmodule